Move FileId type to types
