+++ frontend/src/components/pageEditor/hooks/usePageDocument.ts
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 import { useFileState } from '../../../contexts/FileContext';
 import { PDFDocument, PDFPage } from '../../../types/pageEditor';
-import { FileId } from '../../../types/fileContext';
+import { FileId } from '../../../types/file';
 
 export interface PageDocumentHook {
   document: PDFDocument | null;
